ratings are now saved in database and calculate the average and save them in video column

+++ common/src/main/java/org/group15/tveely/DTOs/rating/RatingRequest.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 public class RatingRequest {
     private String email;
     private Long videoId;
-    private Long rating;
+    private int rating;
 }
